Refactor: Simplify configs and clean up redundant code
Removed `<ImplicitUsings>`, `<Nullable>`, and `<Configurations>`
from multiple `.csproj` files to simplify project configurations.
Eliminated unused `<ItemGroup>` entries and analyzer-related
`<PackageReference>` dependencies to reduce build complexity.

Refactored several classes to improve readability and maintainability:
- Simplified logic in `StateTaxCache`, `CertificateService`, and
  `FrozenReportService` by removing redundant casts and null checks.
- Streamlined method signatures by removing unused parameters
  (e.g., `CancellationToken` in `ReportRunnerService`).
- Fixed logger type mismatch in `TerminatedEmployeeService`.
- Removed unused methods and fields, such as `_fieldCount` in
  `CsvStringHandler` and `LoadAllArchivedChecksumsAsync` in
  `ChecksumValidationService`.

Updated `Directory.Build.props` to centralize MSBuild properties
by importing a parent configuration. Minor formatting fixes and
SQL query optimizations were applied across the codebase.

These changes aim to streamline the codebase, reduce complexity,
and improve overall maintainability and performance.

+++ src/services/src/Demoulas.ProfitSharing.Services/Certificates/CertificateService.cs
@@ -190,7 +190,7 @@ public sealed class CertificateService : ICertificateService
             var age = (byte)cert.DateOfBirth.Age(calInfo.FiscalEndDate.ToDateTime(TimeOnly.MaxValue));
             if (age < 67)
             {
-                age = (byte)67;
+                age = 67;
             }
 
             annuityRates.TryGetValue(age, out var annuityRate);
